Fetch changes recursively and print
We want to find the underlying commits which triggered a particular
build, even if we have to walk multiple linked pipelines to find them.
This is so we can point out the actual commits which may have caused a
failure.

We implement equals/hashCode for MaterialRevision and related classes
using Eclipse's auto-generation feature, and we use this to avoid
adding changes more than once.