Add Log error if config is invalid
Relates to issue #23:
Makes configuration file errors visible in error log and GoCD server.
Previously the error message was show in GoCD server plugin view but it
didn't show the exception that really caused the error. Now the plugin
logs the whole exception stacktrace to the log which can help to debug
configuration issues.